feat(verify): add std-challenge-13 CStr verification and ValidCStr contract support
- Add std-challenge-13: a faithful, self-contained port of core::ffi::CStr
  over a local `#[repr(transparent)] CStr { inner: [u8] }` DST, with a
  `ValidCStr(inner, inner.len())` struct invariant and the std-aligned
  constructor contracts (from_ptr, from_bytes_with_nul_unchecked, strlen,
  CloneToUninit, Index<RangeFrom>). All 17 functions verify SOUND.
- Support `ValidCStr` as a real contract in the symbolic VM:
  * assert_contract_fact materialises the target as a live, initialized,
    nul-terminated allocation and asserts `alloc.size >= n` from the length
    argument (with a DST field-projection fallback to the reference's own
    provenance);
  * check_valid_cstr recognises nul-terminated allocations and their
    sub_alloc_parent sub-slices, and prefers the explicit length argument
    over the allocation size.
- Model `Iterator::position`/`find` (ReturnOptionSomeScanIndex) and `strlen`
  (ReturnScanLength) so scan results carry their bounds, discharging the
  `from_raw_parts` `ValidNum(size_of(T)*len <= isize::MAX)` for symbolic
  lengths like `strlen(ptr) + 1` / `position(bytes) + 1`.
- Use the real length in the bundled std `ValidCStr` contracts
  (bytes.len() / v.len()) instead of the `1` placeholder.
- Extend the `into_iter_local` matcher to `slice::<impl [T]>::iter`.
- Fix free_list_allocator compile error (`usize <= isize::MAX` type mismatch).

+++ rapx/tests/verify_cases/free_list_allocator/src/lib.rs
@@ -64,7 +64,7 @@ impl FreeListAllocator {
     #[rapx::verify]
     pub fn new(size: usize) -> Self {
         assert!(size >= mem::size_of::<FreeBlock>());
-        assert!(size <= isize::MAX);
+        assert!(size <= isize::MAX as usize);
 
         let mut buf = vec![0u8; size];
         let heap = NonNull::new(buf.as_mut_ptr()).expect("non-null after vec alloc");
